easi--print: handle displaying buffers, generalise
Add `presenter' argument. It would be possible to get the presenter in
the body of the function, but it's always a known value whenever
`easi--print' is called, so it's more efficient to call it then.

Handle which buffer to print into, and also handle displaying/hiding
all the relevant buffers.