move presenter code to separate file
Separate the code for the lower-level `presenter' type and its
operations from that of the higher-level `presentable' pseudo-type.

This is a Good Idea because the two are actually quite separate, and
in future there will be a standardised API for working with
presenters, which cannot be accessed from code designed for
presentables.